UserInterface
this has some activities and buttons to navigate around and method
headers for buttons that do stuff.
there is lots of comments and TODO to help clarify.
Im not done trying to get some of the activities working.
I would like to figure out the camera scanning, QR generating, and
google maps API.
But feel free to work on whatever you have time for.